BE/refactor: 도메인이름 login -> auth 변경
