+++ server/src/main/java/com/jh/chamgyunin/domain/auth/dto/userinfo/KakaoUserInfo.java
@@ -1,9 +1,8 @@
-package com.jh.chamgyunin.domain.login.dto.userinfo;
+package com.jh.chamgyunin.domain.auth.dto.userinfo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.ToString;
-import org.springframework.stereotype.Component;
 
 @Getter
 @ToString
