+++ server/src/main/java/com/jh/chamgyunin/domain/auth/exception/InvalidUserProviderException.java
@@ -1,6 +1,6 @@
-package com.jh.chamgyunin.domain.login.exception;
+package com.jh.chamgyunin.domain.auth.exception;
 
-import com.jh.chamgyunin.domain.login.dto.UserProvider;
+import com.jh.chamgyunin.domain.auth.dto.UserProvider;
 import com.jh.chamgyunin.global.error.exception.BusinessException;
 import com.jh.chamgyunin.global.error.exception.ErrorCode;
