+++ server/src/main/java/com/jh/chamgyunin/domain/auth/service/SocialUserInfoService.java
@@ -1,0 +1,7 @@
+package com.jh.chamgyunin.domain.auth.service;
+
+import com.jh.chamgyunin.domain.auth.dto.userinfo.SocialUserInfo;
+
+public interface SocialUserInfoService {
+    SocialUserInfo getUserInfo(final String accessToken);
+}
