--- server/src/main/java/com/jh/chamgyunin/domain/login/service/SocialUserInfoService.java
@@ -1,7 +1,0 @@
-package com.jh.chamgyunin.domain.login.service;
-
-import com.jh.chamgyunin.domain.login.dto.userinfo.SocialUserInfo;
-
-public interface SocialUserInfoService {
-    SocialUserInfo getUserInfo(final String accessToken);
-}
