+++ server/src/main/java/com/jh/chamgyunin/domain/user/dto/SignUpResponse.java
@@ -1,6 +1,6 @@
 package com.jh.chamgyunin.domain.user.dto;
 
-import com.jh.chamgyunin.domain.login.dto.UserProvider;
+import com.jh.chamgyunin.domain.auth.dto.UserProvider;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
