+++ server/src/test/java/com/jh/chamgyunin/domain/user/api/UserControllerTest.java
@@ -1,7 +1,7 @@
 package com.jh.chamgyunin.domain.user.api;
 
 import com.jh.chamgyunin.IntergrationTest;
-import com.jh.chamgyunin.domain.login.dto.UserProvider;
+import com.jh.chamgyunin.domain.auth.dto.UserProvider;
 import com.jh.chamgyunin.domain.user.dao.UserRepository;
 import com.jh.chamgyunin.domain.user.dto.SignUpRequest;
 import com.jh.chamgyunin.domain.user.model.User;
